correction cs/stan/cbf

+++ src/Factory/MatchesFactory.php
@@ -11,7 +11,7 @@ use Nette\Utils\DateTime;
 class MatchesFactory
 {
     /**
-     * @param array $donnees
+     * @param array<string,mixed> $donnees
      * @param Teams $equipe1
      * @param Teams $equipe2
      * @param int $tv1
@@ -49,8 +49,8 @@ class MatchesFactory
             $match->setFStade($stade);
             $match->setDateCreated($dateMatch);
             $match->setStadeAcceuil($donnees['stadeAccueil']);
-            $match->setDepense1(-$donnees['depense1']);
-            $match->setDepense2(-$donnees['depense2']);
+            $match->setDepense1((int) -$donnees['depense1']);
+            $match->setDepense2((int) -$donnees['depense2']);
         }
 
         return $match;
